requirements.txt
pyannote.audio==4.*
huggingface_hub==0.35.*
torch==2.9.*
torchaudio==2.9.*
faster-whisper==1.2.*
ffmpeg-python==0.2.0
python-dotenv==1.1.*